Changed alerts to be a component
